refacto spec
