+++ spec/rails_helper.rb
@@ -37,6 +37,7 @@ RSpec.configure do |config|
   # config.fixture_path = "#{::Rails.root}/spec/fixtures"
   config.include PageSpecHelper
   config.include SigninSpecHelper
+  config.include Select2SpecHelper
   config.include ActiveSupport::Testing::TimeHelpers
   config.include Devise::Test::ControllerHelpers, type: :controller
   config.include Warden::Test::Helpers, type: :feature
